Refactor DownloadService to use a dao for map creation

+++ app/src/main/java/com/peterlaurence/trekme/service/event/MapDownloadEvent.kt
@@ -3,7 +3,7 @@ package com.peterlaurence.trekme.service.event;
 import java.util.*
 
 sealed class MapDownloadEvent
-data class MapDownloadPending(var progress: Double = 100.0): MapDownloadEvent()
+data class MapDownloadPending(var progress: Int = 100): MapDownloadEvent()
 data class MapDownloadFinished(val mapId: UUID): MapDownloadEvent()
 object MapDownloadStorageError: MapDownloadEvent()
 object MapDownloadAlreadyRunning: MapDownloadEvent()
